recover basic functionalities, wait to update 1. fix out of boundary 2. fix plot 3. fix 404 error of utils.js 4.recover logic of node2vec

+++ app.py
@@ -25,10 +25,12 @@ app = Flask(__name__)
 # Configuration for the file upload folder and allowed file types
 UPLOAD_FOLDER = 'uploads'
 RAW_DATA_FOLDER = 'raw_data'
+PROCESSED_GRAPH_FOLDER = 'processed_graph'
 ALLOWED_EXTENSIONS = {'csv', 'xlsx'}
 EPOCHES = 300
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.config['RAW_DATA_FOLDER'] = RAW_DATA_FOLDER
+app.config['PROCESSED_GRAPH_FOLDER'] = PROCESSED_GRAPH_FOLDER
 app.secret_key = 'BabaYaga'
 logging.basicConfig(level=logging.DEBUG,
                     format='%(asctime)s - %(levelname)s - %(message)s')
@@ -40,6 +42,7 @@ def allowed_file(filename):
 
 @app.route('/')
 def home():
+    session.clear()
     return render_template('Home.html')
 
 @app.route('/user_data_adoper', methods=['GET'])
@@ -61,10 +64,13 @@ def upload_data_store():
         node_filepath = session.get('node_filepath')
         edge_filepath = session.get('edge_filepath', None)
         
-        # processor = DataProcessor(node_filepath, edge_filepath) 
+        # def clear_flashed_messages(category_filter=['error']):
+        #     session.pop('_flashes', None)
+        
         column_validation_success, message = column_validation( node_filepath, edge_filepath)
         
         if not column_validation_success:
+            # clear_flashed_messages()
             flash(f"Error: {message}", 'error') 
             return redirect(url_for('upload_user_data'))
 
@@ -100,22 +106,31 @@ def perform_inference():
 
 @app.route('/data_panel')
 def data_process_panel():
-    label_column = session.get('selected_label_column').lower()
-    edge_infer_column = session.get('selected_edge_infer_column').lower()
-    infer_required = session.get('infer_required') == 'True'
-    
+    # initial data
     graph_data = None
+    unique_label = None
+    edge_infer_column = None
     num_features = 0
-    num_classes = 0
+    num_infers = 0
+    num_labels = 0
+    
+    label_column = session.get('selected_label_column').lower()
+    if session.get('selected_edge_infer_column'):
+        edge_infer_column = session.get('selected_edge_infer_column').lower()
+
+    infer_required = session.get('infer_required') == 'True'
     try:
         node_filepath = session.get('node_filepath')
         edge_filepath = session.get('edge_filepath')
         
         if node_filepath:
-            flash("Upload Status: upload successful!")
+            if edge_filepath:
+                flash("Upload Status: upload successful!")
+            else: 
+                flash("Upload Status: node file upload successful!")   
         else:
             logging.error("Upload Status: Sorry, there is something wrong with uploading...")
-        
+
         processor = DataProcessor(
             node_filepath, edge_filepath if edge_filepath else None)
         
@@ -130,10 +145,19 @@ def data_process_panel():
             nodes_data, processor.COLUMN_ALIGNMENT)
         # store index map
         index_to_name_mapping = processor.create_index_id_name_mapping(nodes_data, edge_infer_column)
+        # save index map to processed graph
+        try:
+            mapping_path = os.path.join(PROCESSED_GRAPH_FOLDER, 'edge_mapping.csv')
+            index_to_name_mapping.to_csv(mapping_path, index=False)
+            session['edge_mapping_file'] = mapping_path
+        except Exception as message:
+            logging.error(f'Error: output {str(message)}')
+            flash(f'Error: output {str(message)}')
+
+        if edge_infer_column:
+            unique_infer_categories = nodes_data[edge_infer_column].unique()
         
-        unique_infer_indicator = nodes_data[edge_infer_column].unique()
-        unqiue_label = nodes_data[label_column].unique()
-        # # store a list of all given label
+        unique_label = nodes_data[label_column].unique().tolist()
         
         # prepare num_features and num_classes for visualization of validation
         if label_column == edge_infer_column:
@@ -144,14 +168,16 @@ def data_process_panel():
             col for col in nodes_data.columns if col not in columns_to_exclude]
         
         # column numbers and category of infer indicator
-        num_features, num_labels, num_infers = len(node_features), len(unqiue_label), len(unique_infer_indicator)
+        num_features, num_labels= len(node_features), len(unique_label)
 
         # generate edges
         edges_data = processor.edges_generator(nodes_data, edge_infer_column, edge_filepath)
+
         edge_index = processor.edge_index_generator(edges_data)
         
         # generate features
         x, labels = processor.numeric_dataset(nodes_data, node_features, label_column)
+        
         # check if nan value exists
         flash(processor.nanCheck(nodes_data, x))
         
@@ -161,14 +187,17 @@ def data_process_panel():
         graph_data = processor.construct_graph_data(num_rows, edge_index, x, labels)
         
         # save graph data
-        graph_data_file = 'processed_graph/graph_data.pt'
+        graph_data_file = PROCESSED_GRAPH_FOLDER + '/graph_data.pt'
         torch.save(graph_data, graph_data_file)
         session['graph_data_file'] = graph_data_file
         session['num_features'] = num_features
         session['num_labels'] = num_labels
-        session['label_names'] = unqiue_label.tolist()
+        session["num_infers"] = num_infers
+        session['label_names'] = unique_label
+        session['enable_process'] = False
+        session['enable_download'] = False
+        session['enable_analyze'] = False
     except Exception as e:
-        session['process_success'] = False
         flash(f'Error: {str(e)}')
     finally:
         # Clear the session after processing is complete
@@ -176,69 +205,96 @@ def data_process_panel():
         session.pop('edge_filepath', None)
     
     return render_template('dataProcess.html', 
-                           graph = graph_data,
+                           graph_file = graph_data_file,
+                           mapping_file =mapping_path,
                            num_features=num_features,
                            num_labels=num_labels,
-                           label_names=unqiue_label.tolist())
+                           num_infers=num_infers,
+                           label_names=unique_label,
+                           process_success=session.get('enable_process', False),
+                           enable_download=session.get('enable_download', False),
+                            enable_analyze=session.get('enable_analyze', False))    
+                           
     
 # Implementation of graphSAGE
 @app.route('/process_graphsage', methods=['GET', 'POST'])
 def process_with_graphsage():
+    
     graph_data_file = session.get('graph_data_file')
+    mapping_file = session.get('edge_mapping_file')
     num_features = session.get('num_features')
     num_labels = session.get('num_labels')
+    num_infers = session.get('num_infers')
     labels = session.get('label_names')
-        
+
     if not graph_data_file or num_features is None or num_labels is None:
         logging.error('Missing data for processing.')
         return redirect(url_for('data_process_panel'))
+    if not mapping_file:
+        logging.error('Missing mapping file for processing.')
+        return redirect(url_for('data_process_panel'))
+    
     try:
         graph_data = torch.load(graph_data_file)
         flash('GraphSAGE processing completed successfully', 'success')
+        mapping_df = pd.read_csv(mapping_file)
+        
         os.remove(graph_data_file)
+        os.remove(mapping_file)
         session.pop('graph_data_file', None)
         session.pop('num_features', None)
         session.pop('num_classes', None)
+        session.pop('num_infers', None)
         session.pop('label_names', None)
+        session.pop('edge_mapping_file', None)
         
         graphSAGEProcessor = GraphSAGE(in_channels=num_features, hidden_channels=16, out_channels=num_labels) 
         
         embeddings = graphSAGEProcessor.model_training(graphSAGEProcessor, graph_data, EPOCHES)
+        if embeddings is None:
+            logging.error('Error in generating embeddings.')
+            return redirect(url_for('data_process_panel'))
+        
         # generating validation plot
-        graphSAGEProcessor.generate_tsne_plot(embeddings, labels)
+        edge_embeddings_start = embeddings[graph_data.edge_index[0]]
+        edge_embeddings_end = embeddings[graph_data.edge_index[1]]
         
-        # edge_embeddings_start = embeddings[edge_index[0]]
-        # edge_embeddings_end = embeddings[edge_index[1]]
-        
-        # raw_weights = torch.norm(edge_embeddings_start - edge_embeddings_end, dim=1).cpu().numpy()
-        # edges_with_weights = pd.DataFrame(edge_index.t().cpu().numpy(), columns=['Source', 'Target'])
-        # edges_with_weights['Weight'] = raw_weights
-        
-        # index_to_name_dict = index_to_name_mapping.set_index('index')['name'].to_dict()
-        
-        # edges_with_weights['Source'] = edges_with_weights['Source'].map(index_to_name_dict)
-        # edges_with_weights['Target'] = edges_with_weights['Target'].map(index_to_name_dict)
+        raw_weights = torch.norm(edge_embeddings_start - edge_embeddings_end, dim=1).cpu().numpy()
+        edges_with_weights = pd.DataFrame(graph_data.edge_index.t().cpu().numpy(), columns=['Source', 'Target'])
+        edges_with_weights['Weight'] = raw_weights
+
+        index_to_name_dict = mapping_df.set_index('index')['name'].to_dict()
+
+        edges_with_weights['Source'] = edges_with_weights['Source'].map(index_to_name_dict)
+        edges_with_weights['Target'] = edges_with_weights['Target'].map(index_to_name_dict)
 
         # Save the DataFrame to a CSV file
-        # try:
-        #     output_path = UPLOAD_FOLDER + '/weighted_graph.csv'
-        #     edges_with_weights.to_csv(output_path, index=False)
-        #     message = "Process Status: Congragulations! You data has successfully processed."
-        # except Exception as message:
-        #     flash(f'Error: output {str(message)}')
-        # finally:
-        #     flash(message)
-        #     session['process_success'] = True
-        #     session['data_processed'] = True
-        #     session['processed_file'] = 'weighted_graph.csv'
+        try:
+            output_path = UPLOAD_FOLDER + '/weighted_graph.csv'
+            edges_with_weights.to_csv(output_path, index=False)
+            message = "Process Status: Congragulations! You data has successfully processed."
+        except Exception as message:
+            flash(f'Error: output {str(message)}')
+        finally:
+            flash(message)
+            session['enable_process'] = True
+            session['enable_download'] = True
+            session['enable_analyze'] = True
+            session['processed_file'] = 'weighted_graph.csv'
     except Exception as e:
-        session['process_success'] = False
+        session['enable_process'] = False
+        session['enable_download'] = False
+        session['enable_analyze'] = False
         flash(f'Error: {str(e)}')
+        # return redirect(url_for('graph_infer'))
     finally:
         # Clear the session after processing is complete
         session.pop('node_filepath', None)
         session.pop('edge_filepath', None)
-    return render_template('dataProcess.html', process_success=session.get('process_success', False))
+    return render_template('dataProcess.html', process_success=session.get('enable_process', True),
+                            download_success=session.get('enable_download', True),
+                            analyze_success=session.get('enable_analyze', True))   
+                           
 
 @app.route('/process_node2vec')
 def data_process_node2vec():
@@ -325,7 +381,7 @@ def data_process_node2vec():
         session.pop('node_filepath', None)
         session.pop('edge_filepath', None)
 
-    return render_template('dataProcess.html', process_success=session.get('process_success', False), plot = False)
+    return render_template('dataProcess.html', process_success=session.get('process_success', False))
 
 @app.route('/download_processed_file')
 def download_processed_file():
@@ -568,5 +624,7 @@ if __name__ == '__main__':
         os.makedirs(UPLOAD_FOLDER)
     if not os.path.exists(RAW_DATA_FOLDER):
         os.makedirs(RAW_DATA_FOLDER)
+    if not os.path.exists(PROCESSED_GRAPH_FOLDER):
+        os.makedirs(PROCESSED_GRAPH_FOLDER)
 
     app.run(debug=True)
